Progress bar can be enabled or disabled for masking and labeling processes

+++ src/utils/datasets/_zarrbased.py
@@ -785,5 +785,5 @@ class WeaklyLabeledZarrDataset(DenselyLabeledZarrDataset):
         if self._target_transform:
             target = self._target_transform(target)
 
-        target = target.squeeze()
+        target = target.squeeze().astype(np.int64)
         return patch, target
